Move the ReflectionMethodBodyScanner outside of MarkStep
