+++ src/linker/Linker.Steps/MarkStep.cs
@@ -50,13 +50,13 @@ namespace Mono.Linker.Steps {
 		protected List<MethodBody> _unreachableBodies;
 
 #if DEBUG
-		static DependencyKind [] _entireTypeReasons = new DependencyKind [] {
+		static readonly DependencyKind [] _entireTypeReasons = new DependencyKind [] {
 			DependencyKind.NestedType,
 			DependencyKind.PreservedDependency,
 			DependencyKind.TypeInAssembly,
 		};
 
-		static DependencyKind [] _fieldReasons = new DependencyKind [] {
+		static readonly DependencyKind [] _fieldReasons = new DependencyKind [] {
 			DependencyKind.AccessedViaReflection,
 			DependencyKind.AlreadyMarked,
 			DependencyKind.Custom,
@@ -72,7 +72,7 @@ namespace Mono.Linker.Steps {
 			DependencyKind.TypePreserve,
 		};
 
-		static DependencyKind [] _typeReasons = new DependencyKind [] {
+		static readonly DependencyKind [] _typeReasons = new DependencyKind [] {
 			DependencyKind.AccessedViaReflection,
 			DependencyKind.AlreadyMarked,
 			DependencyKind.AttributeType,
@@ -98,7 +98,7 @@ namespace Mono.Linker.Steps {
 			DependencyKind.VariableType,
 		};
 
-		static DependencyKind [] _methodReasons = new DependencyKind [] {
+		static readonly DependencyKind [] _methodReasons = new DependencyKind [] {
 			DependencyKind.AccessedViaReflection,
 			DependencyKind.AlreadyMarked,
 			DependencyKind.AttributeConstructor,
@@ -592,7 +592,7 @@ namespace Mono.Linker.Steps {
 
 			TypeDefinition td;
 			if (args.Count >= 2 && args [1].Value is string typeName) {
-				td = FindType (assembly ?? context.Module.Assembly, typeName);
+				td = (assembly ?? context.Module.Assembly).FindType (typeName);
 
 				if (td == null) {
 					_context.LogMessage (MessageImportance.Low, $"Could not resolve '{typeName}' type dependency");
@@ -629,14 +629,6 @@ namespace Mono.Linker.Steps {
 				return;
 
 			_context.LogMessage (MessageImportance.High, $"Could not resolve dependency member '{member}' declared in type '{td.FullName}'");
-		}
-
-		static TypeDefinition FindType (AssemblyDefinition assembly, string fullName)
-		{
-			fullName = fullName.ToCecilName ();
-
-			var type = assembly.MainModule.GetType (fullName);
-			return type?.Resolve ();
 		}
 
 		bool MarkDependencyMethod (TypeDefinition type, string name, string[] signature, in DependencyInfo reason)
@@ -849,8 +841,8 @@ namespace Mono.Linker.Steps {
 			MarkCustomAttributeArgument (namedArgument.Argument, ca);
 
 			if (property != null && _flowAnnotations.RequiresDataFlowAnalysis (property.SetMethod)) {
-				var scanner = new ReflectionMethodBodyScanner (this);
-				scanner.ProcessAttributeDataflow (property.SetMethod, new Collection<CustomAttributeArgument> { namedArgument.Argument });
+				var scanner = new ReflectionMethodBodyScanner (_context, this, _flowAnnotations);
+				scanner.ProcessAttributeDataflow (property.SetMethod, new List<CustomAttributeArgument> { namedArgument.Argument });
 			}
 		}
 
@@ -887,7 +879,7 @@ namespace Mono.Linker.Steps {
 			MarkCustomAttributeArgument (namedArgument.Argument, ca);
 
 			if (field != null && _flowAnnotations.RequiresDataFlowAnalysis (field)) {
-				var scanner = new ReflectionMethodBodyScanner (this);
+				var scanner = new ReflectionMethodBodyScanner (_context, this, _flowAnnotations);
 				scanner.ProcessAttributeDataflow (field, namedArgument.Argument);
 			}
 		}
@@ -932,7 +924,7 @@ namespace Mono.Linker.Steps {
 
 			var resolvedConstructor = ca.Constructor.Resolve ();
 			if (resolvedConstructor != null && _flowAnnotations.RequiresDataFlowAnalysis (resolvedConstructor)) {
-				var scanner = new ReflectionMethodBodyScanner (this);
+				var scanner = new ReflectionMethodBodyScanner (_context, this, _flowAnnotations);
 				scanner.ProcessAttributeDataflow (resolvedConstructor, ca.ConstructorArguments);
 			}
 		}
@@ -1109,7 +1101,7 @@ namespace Mono.Linker.Steps {
 			return markOccurred;
 		}
 
-		protected void MarkField (FieldReference reference, DependencyInfo reason)
+		internal protected void MarkField (FieldReference reference, DependencyInfo reason)
 		{
 			if (reference.DeclaringType is GenericInstanceType) {
 				Debug.Assert (reason.Kind == DependencyKind.FieldAccess || reason.Kind == DependencyKind.Ldtoken);
@@ -1189,7 +1181,7 @@ namespace Mono.Linker.Steps {
 				MarkMethodsIf (type.Methods, IsSpecialSerializationConstructor, new DependencyInfo (DependencyKind.SerializationMethodForType, type));
 		}
 
-		protected virtual TypeDefinition MarkType (TypeReference reference, DependencyInfo reason)
+		internal protected virtual TypeDefinition MarkType (TypeReference reference, DependencyInfo reason)
 		{
 #if DEBUG
 			if (!_typeReasons.Contains (reason.Kind))
@@ -1446,7 +1438,7 @@ namespace Mono.Linker.Steps {
 			TypeDefinition tdef = null;
 			switch (attribute.ConstructorArguments [0].Value) {
 			case string s:
-				tdef = ResolveFullyQualifiedTypeName (s);
+				tdef = AssemblyUtilities.ResolveFullyQualifiedTypeName (_context, s);
 				break;
 			case TypeReference type:
 				tdef = type.Resolve ();
@@ -1718,7 +1710,7 @@ namespace Mono.Linker.Steps {
 				parameters [1].ParameterType.Name == "StreamingContext";
 		}
 
-		protected bool MarkMethodsIf (Collection<MethodDefinition> methods, Func<MethodDefinition, bool> predicate, in DependencyInfo reason)
+		internal protected bool MarkMethodsIf (Collection<MethodDefinition> methods, Func<MethodDefinition, bool> predicate, in DependencyInfo reason)
 		{
 			bool marked = false;
 			foreach (MethodDefinition method in methods) {
@@ -1826,25 +1818,6 @@ namespace Mono.Linker.Steps {
 		protected virtual void MarkMulticastDelegate (TypeDefinition type)
 		{
 			MarkMethodCollection (type.Methods, new DependencyInfo (DependencyKind.MethodForSpecialType, type));
-		}
-
-		TypeDefinition ResolveFullyQualifiedTypeName (string name)
-		{
-			if (!TypeNameParser.TryParseTypeAssemblyQualifiedName (name, out string typeName, out string assemblyName))
-				return null;
-
-			foreach (var assemblyDefinition in _context.GetAssemblies ()) {
-				if (assemblyName != null && assemblyDefinition.Name.Name != assemblyName)
-					continue;
-
-				var foundType = assemblyDefinition.MainModule.GetType (typeName);
-				if (foundType == null)
-					continue;
-
-				return foundType;
-			}
-
-			return null;
 		}
 
 		protected (TypeReference, DependencyInfo) GetOriginalType (TypeReference type, DependencyInfo reason)
@@ -2045,7 +2018,7 @@ namespace Mono.Linker.Steps {
 				MarkMethod (method, reason);
 		}
 
-		protected void MarkIndirectlyCalledMethod (MethodDefinition method, in DependencyInfo reason)
+		internal protected void MarkIndirectlyCalledMethod (MethodDefinition method, in DependencyInfo reason)
 		{
 			MarkMethod (method, reason);
 			Annotations.MarkIndirectlyCalledMethod (method);
@@ -2396,7 +2369,7 @@ namespace Mono.Linker.Steps {
 			}
 		}
 
-		protected void MarkProperty (PropertyDefinition prop, in DependencyInfo reason)
+		internal protected void MarkProperty (PropertyDefinition prop, in DependencyInfo reason)
 		{
 			Tracer.AddDirectDependency (prop, reason, marked: false);
 			// Consider making this more similar to MarkEvent method?
@@ -2404,7 +2377,7 @@ namespace Mono.Linker.Steps {
 			DoAdditionalPropertyProcessing (prop);
 		}
 
-		protected virtual void MarkEvent (EventDefinition evt, in DependencyInfo reason)
+		internal protected virtual void MarkEvent (EventDefinition evt, in DependencyInfo reason)
 		{
 			// Record the event without marking it in Annotations.
 			Tracer.AddDirectDependency (evt, reason, marked: false);
@@ -2415,7 +2388,7 @@ namespace Mono.Linker.Steps {
 			DoAdditionalEventProcessing (evt);
 		}
 
-		void MarkMethodIfNotNull (MethodReference method, in DependencyInfo reason)
+		internal void MarkMethodIfNotNull (MethodReference method, in DependencyInfo reason)
 		{
 			if (method == null)
 				return;
@@ -2599,7 +2572,7 @@ namespace Mono.Linker.Steps {
 				return;
 
 			if (requiresReflectionMethodBodyScanner) {
-				var scanner = new ReflectionMethodBodyScanner (this);
+				var scanner = new ReflectionMethodBodyScanner (_context, this, _flowAnnotations);
 				scanner.ScanAndProcessReturnValue (body);
 			}
 
@@ -2619,90 +2592,6 @@ namespace Mono.Linker.Steps {
 			}
 		}
 
-		/// <summary>
-		/// Helper struct to pass around context information about reflection pattern
-		/// as a single parameter (and have a way to extend this in the future if we need to easily).
-		/// Also implements a simple validation mechanism to check that the code does report patter recognition
-		/// results for all methods it works on.
-		/// The promise of the pattern recorder is that for a given reflection method, it will either not talk
-		/// about it ever, or it will always report recognized/unrecognized.
-		/// </summary>
-		struct ReflectionPatternContext : IDisposable
-		{
-			readonly LinkContext _context;
-#if DEBUG
-			bool _patternAnalysisAttempted;
-			bool _patternReported;
-#endif
-
-			public MethodDefinition MethodCalling { get; private set; }
-			public MethodDefinition MethodCalled { get; private set; }
-			public int InstructionIndex { get; private set; }
-
-			public ReflectionPatternContext (LinkContext context, MethodDefinition methodCalling, MethodDefinition methodCalled, int instructionIndex)
-			{
-				_context = context;
-				MethodCalling = methodCalling;
-				MethodCalled = methodCalled;
-				InstructionIndex = instructionIndex;
-
-#if DEBUG
-				_patternAnalysisAttempted = false;
-				_patternReported = false;
-#endif
-			}
-
-			[Conditional("DEBUG")]
-			public void AnalyzingPattern ()
-			{
-#if DEBUG
-				_patternAnalysisAttempted = true;
-#endif
-			}
-
-			[Conditional ("DEBUG")]
-			public void RecordHandledPattern ()
-			{
-#if DEBUG
-				_patternReported = true;
-#endif
-			}
-
-			public void RecordRecognizedPattern<T> (T accessedItem, Action mark)
-				where T : IMemberDefinition
-			{
-#if DEBUG
-				if (!_patternAnalysisAttempted)
-					throw new InvalidOperationException ($"Internal error: To correctly report all patterns, when starting to analyze a pattern the AnalyzingPattern must be called first. {MethodCalling} -> {MethodCalled}");
-
-				_patternReported = true;
-#endif
-
-				mark ();
-				_context.ReflectionPatternRecorder.RecognizedReflectionAccessPattern (MethodCalling, MethodCalled, accessedItem);
-			}
-
-			public void RecordUnrecognizedPattern (string message)
-			{
-#if DEBUG
-				if (!_patternAnalysisAttempted)
-					throw new InvalidOperationException ($"Internal error: To correctly report all patterns, when starting to analyze a pattern the AnalyzingPattern must be called first. {MethodCalling} -> {MethodCalled}");
-
-				_patternReported = true;
-#endif
-
-				_context.ReflectionPatternRecorder.UnrecognizedReflectionAccessPattern (MethodCalling, MethodCalled, message);
-			}
-
-			public void Dispose ()
-			{
-#if DEBUG
-				if (_patternAnalysisAttempted && !_patternReported)
-					throw new InvalidOperationException ($"Internal error: A reflection pattern was analyzed, but no result was reported. {MethodCalling} -> {MethodCalled}");
-#endif
-			}
-		}
-
 		protected class AttributeProviderPair {
 			public AttributeProviderPair (CustomAttribute attribute, ICustomAttributeProvider provider)
 			{
@@ -2712,1494 +2601,6 @@ namespace Mono.Linker.Steps {
 
 			public CustomAttribute Attribute { get; private set; }
 			public ICustomAttributeProvider Provider { get; private set; }
-		}
-
-		private class ReflectionMethodBodyScanner : Dataflow.MethodBodyScanner
-		{
-			private readonly MarkStep _markStep;
-			private readonly FlowAnnotations _flowAnnotations;
-
-			public static bool RequiresReflectionMethodBodyScanner (FlowAnnotations flowAnnotations, MethodReference method)
-			{
-				MethodDefinition methodDefinition = method.Resolve ();
-				if (methodDefinition != null) {
-					return GetIntrinsicIdForMethod (methodDefinition) > IntrinsicId.RequiresReflectionBodyScanner_Sentinel || flowAnnotations.RequiresDataFlowAnalysis (methodDefinition);
-				}
-
-				return false;
-			}
-
-			public static bool RequiresReflectionMethodBodyScanner (FlowAnnotations flowAnnotations, FieldReference field)
-			{
-				FieldDefinition fieldDefinition = field.Resolve ();
-				if (fieldDefinition != null)
-					return flowAnnotations.RequiresDataFlowAnalysis (fieldDefinition);
-
-				return false;
-			}
-
-			public ReflectionMethodBodyScanner(MarkStep parent)
-			{
-				_markStep = parent;
-				_flowAnnotations = _markStep._flowAnnotations;
-			}
-
-			public void ScanAndProcessReturnValue (MethodBody methodBody)
-			{
-				Scan (methodBody);
-
-				if (MethodReturnValue != null) {
-					var requiredMemberKinds = _flowAnnotations.GetReturnParameterAnnotation (methodBody.Method);
-					if (requiredMemberKinds != 0) {
-						var reflectionContext = new ReflectionPatternContext (_markStep._context, methodBody.Method, methodBody.Method, 0);
-						reflectionContext.AnalyzingPattern ();
-						RequireDynamicallyAccessedMembers (ref reflectionContext, requiredMemberKinds, MethodReturnValue, methodBody.Method.MethodReturnType);
-					}
-				}
-			}
-
-			public void ProcessAttributeDataflow (MethodDefinition method, Collection<CustomAttributeArgument> arguments)
-			{
-				int paramOffset = method.HasImplicitThis () ? 1 : 0;
-
-				for (int i = 0; i < method.Parameters.Count; i++)
-				{
-					var annotation = _flowAnnotations.GetParameterAnnotation (method, i + paramOffset);
-					if (annotation != 0) {
-						ValueNode valueNode = GetValueNodeForCustomAttributeArgument (arguments [i]);
-						if (valueNode != null) {
-							// TODO: There's no way to represent attribute dataflow in current ReflectionPatternContext (and the underlying IReflectionPatterRecorder)
-							ReflectionPatternContext context = new ReflectionPatternContext (_markStep._context, method, method, 0);
-							context.AnalyzingPattern ();
-							RequireDynamicallyAccessedMembers (ref context, annotation, valueNode, method);
-						}
-					}
-				}
-			}
-
-			public void ProcessAttributeDataflow (FieldDefinition field, CustomAttributeArgument value)
-			{
-				var annotation = _flowAnnotations.GetFieldAnnotation (field);
-				Debug.Assert (annotation != 0);
-
-				ValueNode valueNode = GetValueNodeForCustomAttributeArgument (value);
-				if (valueNode != null) {
-					// TODO: There's no way to represent store to a field given current ReflectionPatternContext (and the underlying IReflectionPatterRecorder)
-					var reflectionContext = new ReflectionPatternContext (_markStep._context, field.DeclaringType.Methods [0], field.DeclaringType.Methods [0], 0);
-					reflectionContext.AnalyzingPattern ();
-					RequireDynamicallyAccessedMembers (ref reflectionContext, annotation, valueNode, field);
-				}
-			}
-
-			private ValueNode GetValueNodeForCustomAttributeArgument(CustomAttributeArgument argument)
-			{
-				ValueNode valueNode;
-				if (argument.Type.Name == "Type") {
-					TypeDefinition referencedType = ((TypeReference)argument.Value).Resolve ();
-					valueNode = referencedType == null ? null : new SystemTypeValue (referencedType);
-				} else if (argument.Type.MetadataType == MetadataType.String) {
-					valueNode = new KnownStringValue ((string)argument.Value);
-				} else {
-					// We shouldn't have gotten a non-null annotation for this from GetParameterAnnotation
-					throw new InvalidOperationException ();
-				}
-
-				return valueNode;
-			}
-
-			protected override void WarnAboutInvalidILInMethod (MethodBody method, int ilOffset)
-			{
-				// TODO: remove once we're ready to scan actual invalid IL
-				// Serves as a debug helper for now to make sure valid IL is not considered invalid.
-				throw new Exception ();
-			}
-
-			protected override ValueNode GetMethodParameterValue (MethodDefinition method, int parameterIndex)
-			{
-				DynamicallyAccessedMemberKinds memberKinds = _flowAnnotations.GetParameterAnnotation (method, parameterIndex);
-				return new MethodParameterValue (parameterIndex, memberKinds) {
-					SourceContext = method
-				};
-			}
-
-			protected override ValueNode GetFieldValue (MethodDefinition method, FieldDefinition field)
-			{
-				DynamicallyAccessedMemberKinds memberKinds = _flowAnnotations.GetFieldAnnotation (field);
-				return new LoadFieldValue (field, memberKinds) {
-					SourceContext = method
-				};
-			}
-
-			protected override void HandleStoreField (MethodDefinition method, FieldDefinition field, Instruction operation, ValueNode valueToStore)
-			{
-				var requiredMemberKinds = _flowAnnotations.GetFieldAnnotation (field);
-				if (requiredMemberKinds != 0) {
-					// TODO: There's no way to represent store to a field given current ReflectionPatternContext (and the underlying IReflectionPatterRecorder)
-					var reflectionContext = new ReflectionPatternContext (_markStep._context, method, method, operation.Offset);
-					reflectionContext.AnalyzingPattern ();
-					RequireDynamicallyAccessedMembers (ref reflectionContext, requiredMemberKinds, valueToStore, field);
-				}
-			}
-
-			protected override void HandleStoreParameter (MethodDefinition method, int index, Instruction operation, ValueNode valueToStore)
-			{
-				var requiredMemberKinds = _flowAnnotations.GetParameterAnnotation (method, index);
-				if (requiredMemberKinds != 0) {
-					// TODO: There's no way to represent store to a parameter given current ReflectionPatternContext (and the underlying IReflectionPatterRecorder)
-					var reflectionContext = new ReflectionPatternContext (_markStep._context, method, method, operation.Offset);
-					reflectionContext.AnalyzingPattern ();
-					RequireDynamicallyAccessedMembers (ref reflectionContext, requiredMemberKinds, valueToStore, method.Parameters[index - (method.HasImplicitThis() ? 1 : 0)]);
-				}
-			}
-
-			enum IntrinsicId
-			{
-				None = 0,
-				IntrospectionExtensions_GetTypeInfo,
-				Type_GetTypeFromHandle,
-
-				// Anything above this marker will require the method to be run through
-				// the reflection body scanner.
-				RequiresReflectionBodyScanner_Sentinel = 1000,
-				Type_MakeGenericType,
-				Type_GetType,
-				Type_GetConstructor,
-				Type_GetMethod,
-				Type_GetField,
-				Type_GetProperty,
-				Type_GetEvent,
-				Expression_Call,
-				Expression_Field,
-				Expression_Property,
-				Expression_New,
-				Activator_CreateInstance_Type,
-				Activator_CreateInstance_AssemblyName_TypeName,
-				Activator_CreateInstanceFrom,
-				Activator_CreateInstanceOfT,
-				AppDomain_CreateInstance,
-				AppDomain_CreateInstanceAndUnwrap,
-				AppDomain_CreateInstanceFrom,
-				AppDomain_CreateInstanceFromAndUnwrap,
-				Assembly_CreateInstance,
-				RuntimeReflectionExtensions_GetRuntimeEvent,
-				RuntimeReflectionExtensions_GetRuntimeField,
-				RuntimeReflectionExtensions_GetRuntimeMethod,
-				RuntimeReflectionExtensions_GetRuntimeProperty
-			}
-
-			static IntrinsicId GetIntrinsicIdForMethod (MethodDefinition calledMethod)
-			{
-				switch (calledMethod.Name) {
-					case "GetTypeInfo" when calledMethod.DeclaringType.Name == "IntrospectionExtensions"
-						&& calledMethod.DeclaringType.Namespace == "System.Reflection":
-						return IntrinsicId.IntrospectionExtensions_GetTypeInfo;
-
-					case "GetTypeFromHandle" when calledMethod.DeclaringType.Name == "Type"
-						&& calledMethod.DeclaringType.Namespace == "System":
-						return IntrinsicId.Type_GetTypeFromHandle;
-
-					case "MakeGenericType" when calledMethod.DeclaringType.Name == "Type"
-						&& calledMethod.DeclaringType.Namespace == "System":
-						return IntrinsicId.Type_MakeGenericType;
-
-					//
-					// System.Reflection.RuntimeReflectionExtensions
-					//
-					// static GetRuntimeEvent (this Type type, string name)
-					// static GetRuntimeField (this Type type, string name)
-					// static GetRuntimeMethod (this Type type, string name, Type[] parameters)
-					// static GetRuntimeProperty (this Type type, string name)
-					//
-					case var getRuntimeMember when (getRuntimeMember.StartsWith ("GetRuntime")
-						&& calledMethod.DeclaringType.Namespace == "System.Reflection"
-						&& calledMethod.DeclaringType.Name == "RuntimeReflectionExtensions"
-						&& calledMethod.Parameters.Count >= 2
-						&& calledMethod.Parameters [0].ParameterType.FullName == "System.Type"
-						&& calledMethod.Parameters [1].ParameterType.FullName == "System.String"):
-						return getRuntimeMember switch
-						{
-							"GetRuntimeEvent" => IntrinsicId.RuntimeReflectionExtensions_GetRuntimeEvent,
-							"GetRuntimeField" => IntrinsicId.RuntimeReflectionExtensions_GetRuntimeField,
-							"GetRuntimeMethod" => IntrinsicId.RuntimeReflectionExtensions_GetRuntimeMethod,
-							"GetRuntimeProperty" => IntrinsicId.RuntimeReflectionExtensions_GetRuntimeProperty,
-							_ => throw new ArgumentException($"Reflection call '{calledMethod.FullName}' is of unexpected member type.")
-						};
-
-					//
-					// System.Linq.Expressions.Expression
-					// 
-					// static Call (Type, String, Type[], Expression[])
-					//
-					case "Call" when calledMethod.DeclaringType.Name == "Expression"
-						&& calledMethod.DeclaringType.Namespace == "System.Linq.Expressions"
-						&& calledMethod.Parameters.Count == 4
-						&& calledMethod.Parameters [0].ParameterType.FullName == "System.Type":
-						return IntrinsicId.Expression_Call;
-
-					//
-					// System.Linq.Expressions.Expression
-					// 
-					// static Field (Expression, Type, String)
-					// static Property (Expression, Type, String)
-					//
-					case var fieldOrProperty when ((fieldOrProperty == "Field" || fieldOrProperty == "Property")
-						&& calledMethod.DeclaringType.Namespace == "System.Linq.Expressions"
-						&& calledMethod.DeclaringType.Name == "Expression"
-						&& calledMethod.Parameters.Count == 3
-						&& calledMethod.Parameters [1].ParameterType.FullName == "System.Type"):
-						return fieldOrProperty [0] == 'P' ? IntrinsicId.Expression_Property : IntrinsicId.Expression_Field;
-
-					//
-					// System.Linq.Expressions.Expression
-					// 
-					// static New (Type)
-					//
-					case "New" when calledMethod.DeclaringType.Name == "Expression"
-						&& calledMethod.DeclaringType.Namespace == "System.Linq.Expressions"
-						&& calledMethod.Parameters.Count == 1
-						&& calledMethod.Parameters [0].ParameterType.FullName == "System.Type":
-						return IntrinsicId.Expression_New;
-
-					//
-					// System.Type
-					// 
-					// GetType (string)
-					// GetType (string, Boolean)
-					// GetType (string, Boolean, Boolean)
-					// GetType (string, Func<AssemblyName, Assembly>, Func<Assembly, String, Boolean, Type>)
-					// GetType (string, Func<AssemblyName, Assembly>, Func<Assembly, String, Boolean, Type>, Boolean)
-					// GetType (string, Func<AssemblyName, Assembly>, Func<Assembly, String, Boolean, Type>, Boolean, Boolean)
-					//
-					case "GetType" when calledMethod.DeclaringType.Name == "Type"
-						&& calledMethod.DeclaringType.Namespace == "System"
-						&& calledMethod.Parameters.Count >= 1
-						&& calledMethod.Parameters [0].ParameterType.MetadataType == MetadataType.String:
-						return IntrinsicId.Type_GetType;
-
-					//
-					// System.Type
-					// 
-					// GetConstructor (Type[])
-					// GetConstructor (BindingFlags, Binder, Type[], ParameterModifier [])
-					// GetConstructor (BindingFlags, Binder, CallingConventions, Type[], ParameterModifier [])
-					//
-					case "GetConstructor" when calledMethod.DeclaringType.Name == "Type"
-						&& calledMethod.DeclaringType.Namespace == "System"
-						&& calledMethod.HasThis:
-						return IntrinsicId.Type_GetConstructor;
-
-					//
-					// System.Type
-					// 
-					// GetMethod (string)
-					// GetMethod (string, BindingFlags)
-					// GetMethod (string, Type[])
-					// GetMethod (string, Type[], ParameterModifier[])
-					// GetMethod (string, BindingFlags, Binder, Type[], ParameterModifier[]) 6
-					// GetMethod (string, BindingFlags, Binder, CallingConventions, Type[], ParameterModifier[]) 7
-					// GetMethod (string, int, Type[])
-					// GetMethod (string, int, Type[], ParameterModifier[]?)
-					// GetMethod (string, int, BindingFlags, Binder?, Type[], ParameterModifier[]?)
-					// GetMethod (string, int, BindingFlags, Binder?, CallingConventions, Type[], ParameterModifier[]?)
-					//
-					case "GetMethod" when calledMethod.DeclaringType.Name == "Type"
-						&& calledMethod.DeclaringType.Namespace == "System"
-						&& calledMethod.Parameters.Count >= 1
-						&& calledMethod.HasThis:
-						return IntrinsicId.Type_GetMethod;
-
-					//
-					// System.Type
-					// 
-					// GetField (string)
-					// GetField (string, BindingFlags)
-					// GetEvent (string)
-					// GetEvent (string, BindingFlags)
-					// GetProperty (string)
-					// GetProperty (string, BindingFlags)
-					// GetProperty (string, Type)
-					// GetProperty (string, Type[])
-					// GetProperty (string, Type, Type[])
-					// GetProperty (string, Type, Type[], ParameterModifier[])
-					// GetProperty (string, BindingFlags, Binder, Type, Type[], ParameterModifier[])
-					//
-					case var fieldPropertyOrEvent when ((fieldPropertyOrEvent == "GetField" || fieldPropertyOrEvent == "GetProperty" || fieldPropertyOrEvent == "GetEvent")
-						&& calledMethod.DeclaringType.Namespace == "System"
-						&& calledMethod.DeclaringType.Name == "Type"
-						&& calledMethod.Parameters [0].ParameterType.FullName == "System.String")
-						&& calledMethod.HasThis:
-						return fieldPropertyOrEvent switch
-						{
-							"GetField" => IntrinsicId.Type_GetField,
-							"GetProperty" => IntrinsicId.Type_GetProperty,
-							"GetEvent" => IntrinsicId.Type_GetEvent,
-							_ => throw new ArgumentException ($"Reflection call '{calledMethod.FullName}' is of unexpected member type.")
-						};
-
-					//
-					// System.Activator
-					// 
-					// static CreateInstance (System.Type type)
-					// static CreateInstance (System.Type type, bool nonPublic)
-					// static CreateInstance (System.Type type, params object?[]? args)
-					// static CreateInstance (System.Type type, object?[]? args, object?[]? activationAttributes)
-					// static CreateInstance (System.Type type, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object?[]? args, System.Globalization.CultureInfo? culture)
-					// static CreateInstance (System.Type type, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object?[]? args, System.Globalization.CultureInfo? culture, object?[]? activationAttributes) { throw null; }
-					//
-					case "CreateInstance" when !calledMethod.ContainsGenericParameter
-						&& calledMethod.DeclaringType.Namespace == "System"
-						&& calledMethod.DeclaringType.Name == "Activator"
-						&& calledMethod.Parameters.Count >= 1
-						&& calledMethod.DeclaringType.Namespace == "System"
-						&& calledMethod.Parameters [0].ParameterType.MetadataType != MetadataType.String:
-						return IntrinsicId.Activator_CreateInstance_Type;
-
-					//
-					// System.Activator
-					// 
-					// static CreateInstance (string assemblyName, string typeName)
-					// static CreateInstance (string assemblyName, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object?[]? args, System.Globalization.CultureInfo? culture, object?[]? activationAttributes)
-					// static CreateInstance (string assemblyName, string typeName, object?[]? activationAttributes)
-					//
-					case "CreateInstance" when !calledMethod.ContainsGenericParameter
-						&& calledMethod.DeclaringType.Namespace == "System"
-						&& calledMethod.DeclaringType.Name == "Activator"
-						&& calledMethod.Parameters.Count >= 2
-						&& calledMethod.Parameters [0].ParameterType.MetadataType == MetadataType.String
-						&& calledMethod.Parameters [1].ParameterType.MetadataType == MetadataType.String:
-						return IntrinsicId.Activator_CreateInstance_AssemblyName_TypeName;
-
-					//
-					// System.Activator
-					// 
-					// static CreateInstanceFrom (string assemblyFile, string typeName)
-					// static CreateInstanceFrom (string assemblyFile, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object? []? args, System.Globalization.CultureInfo? culture, object? []? activationAttributes)
-					// static CreateInstanceFrom (string assemblyFile, string typeName, object? []? activationAttributes)
-					//
-					case "CreateInstanceFrom" when !calledMethod.ContainsGenericParameter
-						&& calledMethod.DeclaringType.Namespace == "System"
-						&& calledMethod.DeclaringType.Name == "Activator"
-						&& calledMethod.Parameters.Count >= 2
-						&& calledMethod.Parameters [0].ParameterType.MetadataType == MetadataType.String
-						&& calledMethod.Parameters [1].ParameterType.MetadataType == MetadataType.String:
-						return IntrinsicId.Activator_CreateInstanceFrom;
-
-					//
-					// System.Activator
-					// 
-					// static T CreateInstance<T> ()
-					//
-					case "CreateInstance" when calledMethod.ContainsGenericParameter
-						&& calledMethod.DeclaringType.Namespace == "System"
-						&& calledMethod.DeclaringType.Name == "Activator"
-						&& calledMethod.Parameters.Count == 0:
-						return IntrinsicId.Activator_CreateInstanceOfT;
-
-					//
-					// System.AppDomain
-					//
-					// CreateInstance (string assemblyName, string typeName)
-					// CreateInstance (string assemblyName, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object? []? args, System.Globalization.CultureInfo? culture, object? []? activationAttributes)
-					// CreateInstance (string assemblyName, string typeName, object? []? activationAttributes)
-					//
-					// CreateInstanceAndUnwrap (string assemblyName, string typeName)
-					// CreateInstanceAndUnwrap (string assemblyName, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object? []? args, System.Globalization.CultureInfo? culture, object? []? activationAttributes)
-					// CreateInstanceAndUnwrap (string assemblyName, string typeName, object? []? activationAttributes)
-					//
-					// CreateInstanceFrom (string assemblyFile, string typeName)
-					// CreateInstanceFrom (string assemblyFile, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object? []? args, System.Globalization.CultureInfo? culture, object? []? activationAttributes)
-					// CreateInstanceFrom (string assemblyFile, string typeName, object? []? activationAttributes)
-					//
-					// CreateInstanceFromAndUnwrap (string assemblyFile, string typeName)
-					// CreateInstanceFromAndUnwrap (string assemblyFile, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object? []? args, System.Globalization.CultureInfo? culture, object? []? activationAttributes)
-					// CreateInstanceFromAndUnwrap (string assemblyFile, string typeName, object? []? activationAttributes)
-					//
-					case var appDomainCreateInstanceMethodName when calledMethod.DeclaringType.Name == "AppDomain" && calledMethod.DeclaringType.Namespace == "System"
-						&& (appDomainCreateInstanceMethodName == "CreateInstance" || appDomainCreateInstanceMethodName == "CreateInstanceAndUnwrap"
-							|| appDomainCreateInstanceMethodName == "CreateInstanceFrom" || appDomainCreateInstanceMethodName == "CreateInstanceFromAndUnwrap")
-						&& calledMethod.Parameters.Count >= 2
-						&& calledMethod.Parameters [0].ParameterType.MetadataType == MetadataType.String
-						&& calledMethod.Parameters [1].ParameterType.MetadataType == MetadataType.String:
-						return appDomainCreateInstanceMethodName switch
-						{
-							"CreateInstance" => IntrinsicId.AppDomain_CreateInstance,
-							"CreateInstanceAndUnwrap" => IntrinsicId.AppDomain_CreateInstanceAndUnwrap,
-							"CreateInstanceFrom" => IntrinsicId.AppDomain_CreateInstanceFrom,
-							"CreateInstanceFromAndUnwrap" => IntrinsicId.AppDomain_CreateInstanceFromAndUnwrap,
-							_ => throw new ArgumentException ($"Reflection call '{calledMethod.FullName}' is unexpected.")
-						};
-
-					//
-					// System.Reflection.Assembly
-					//
-					// CreateInstance (string typeName)
-					// CreateInstance (string typeName, bool ignoreCase)
-					// CreateInstance (string typeName, bool ignoreCase, BindingFlags bindingAttr, Binder? binder, object []? args, CultureInfo? culture, object []? activationAttributes)
-					//
-					case "CreateInstance" when calledMethod.DeclaringType.Name == "Assembly" && calledMethod.DeclaringType.Namespace == "System.Reflection"
-						&& calledMethod.Parameters.Count >= 1
-						&& calledMethod.Parameters [0].ParameterType.MetadataType == MetadataType.String:
-						return IntrinsicId.Assembly_CreateInstance;
-
-					default:
-						return IntrinsicId.None;
-				}
-			}
-
-			public override bool HandleCall (MethodBody callingMethodBody, MethodReference calledMethod, Instruction operation, ValueNodeList methodParams, out ValueNode methodReturnValue)
-			{
-				var reflectionContext = new ReflectionPatternContext (_markStep._context, callingMethodBody.Method, calledMethod.Resolve (), operation.Offset);
-
-				DynamicallyAccessedMemberKinds returnValueDynamicallyAccessedMemberKinds = 0;
-
-				methodReturnValue = null;
-
-				var calledMethodDefinition = calledMethod.Resolve ();
-				if (calledMethodDefinition == null)
-					return false;
-
-				try {
-
-
-					bool requiresDataFlowAnalysis = _flowAnnotations.RequiresDataFlowAnalysis (calledMethodDefinition);
-					returnValueDynamicallyAccessedMemberKinds =  requiresDataFlowAnalysis ?
-						_flowAnnotations.GetReturnParameterAnnotation (calledMethodDefinition) : 0;
-
-					switch (GetIntrinsicIdForMethod (calledMethodDefinition)) {
-						case IntrinsicId.IntrospectionExtensions_GetTypeInfo: {
-								// typeof(Foo).GetTypeInfo()... will be commonly present in code targeting
-								// the dead-end reflection refactoring. The call doesn't do anything and we
-								// don't want to lose the annotation.
-								methodReturnValue = methodParams [0];
-							}
-							break;
-
-						case IntrinsicId.Type_GetTypeFromHandle: {
-								// Infrastructure piece to support "typeof(Foo)"
-								if (methodParams [0] is RuntimeTypeHandleValue typeHandle)
-									methodReturnValue = new SystemTypeValue (typeHandle.TypeRepresented);
-							}
-							break;
-
-						case IntrinsicId.Type_MakeGenericType: {
-								// Don't care about the actual arguments, but we don't want to lose track of the type
-								// in case this is e.g. Activator.CreateInstance(typeof(Foo<>).MakeGenericType(...));
-								methodReturnValue = methodParams [0];
-							}
-							break;
-
-						//
-						// System.Reflection.RuntimeReflectionExtensions
-						//
-						// static GetRuntimeEvent (this Type type, string name)
-						// static GetRuntimeField (this Type type, string name)
-						// static GetRuntimeMethod (this Type type, string name, Type[] parameters)
-						// static GetRuntimeProperty (this Type type, string name)
-						//
-						case var getRuntimeMember when getRuntimeMember == IntrinsicId.RuntimeReflectionExtensions_GetRuntimeEvent
-							|| getRuntimeMember == IntrinsicId.RuntimeReflectionExtensions_GetRuntimeField
-							|| getRuntimeMember == IntrinsicId.RuntimeReflectionExtensions_GetRuntimeMethod
-							|| getRuntimeMember == IntrinsicId.RuntimeReflectionExtensions_GetRuntimeProperty: {
-
-								reflectionContext.AnalyzingPattern ();
-								DynamicallyAccessedMemberKinds? memberKind = null;
-								BindingFlags bindingFlags = BindingFlags.Instance | BindingFlags.Static | BindingFlags.Public;
-
-								switch (getRuntimeMember) {
-									case IntrinsicId.RuntimeReflectionExtensions_GetRuntimeEvent:
-										memberKind = DynamicallyAccessedMemberKinds.PublicEvents;
-										break;
-
-									case IntrinsicId.RuntimeReflectionExtensions_GetRuntimeField:
-										memberKind = DynamicallyAccessedMemberKinds.PublicFields;
-										break;
-
-									case IntrinsicId.RuntimeReflectionExtensions_GetRuntimeMethod:
-										memberKind = DynamicallyAccessedMemberKinds.PublicMethods;
-										break;
-
-									case IntrinsicId.RuntimeReflectionExtensions_GetRuntimeProperty:
-										memberKind = DynamicallyAccessedMemberKinds.PublicProperties;
-										break;
-
-									default:
-										Debug.Fail ("Unsupported member type");
-										reflectionContext.RecordUnrecognizedPattern ($"Reflection call '{calledMethod.FullName}' inside '{callingMethodBody.Method.FullName}' is of unexpected member type.");
-										break;
-								}
-
-								if (memberKind == null)
-									break;
-
-								foreach (var value in methodParams [0].UniqueValues ()) {
-									if (value is SystemTypeValue systemTypeValue) {
-										foreach (var stringParam in methodParams [1].UniqueValues ()) {
-											if (stringParam is KnownStringValue stringValue) {
-												switch (memberKind) {
-													case DynamicallyAccessedMemberKinds.PublicEvents:
-														MarkEventsOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, e => e.Name == stringValue.Contents, bindingFlags);
-														reflectionContext.RecordHandledPattern ();
-														break;
-													case DynamicallyAccessedMemberKinds.PublicFields:
-														MarkFieldsOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, f => f.Name == stringValue.Contents, bindingFlags);
-														reflectionContext.RecordHandledPattern ();
-														break;
-													case DynamicallyAccessedMemberKinds.PublicMethods:
-														MarkMethodsOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, m => m.Name == stringValue.Contents, bindingFlags);
-														reflectionContext.RecordHandledPattern ();
-														break;
-													case DynamicallyAccessedMemberKinds.PublicProperties:
-														MarkPropertiesOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, p => p.Name == stringValue.Contents, bindingFlags);
-														reflectionContext.RecordHandledPattern ();
-														break;
-													default:
-														throw new MarkException (string.Format ("Error processing reflection call: '{0}' inside {1}. Unexpected member kind.", calledMethod.FullName, callingMethodBody.Method.FullName));
-												}
-											} else {
-												RequireDynamicallyAccessedMembers (ref reflectionContext, (DynamicallyAccessedMemberKinds)memberKind, value, calledMethod.Parameters [0]);
-											}
-										}
-									} else {
-										RequireDynamicallyAccessedMembers (ref reflectionContext, (DynamicallyAccessedMemberKinds)memberKind, value, calledMethod.Parameters [0]);
-									}
-								}
-							}
-							break;
-
-						//
-						// System.Linq.Expressions.Expression
-						// 
-						// static Call (Type, String, Type[], Expression[])
-						//
-						case IntrinsicId.Expression_Call: {
-								reflectionContext.AnalyzingPattern ();
-								BindingFlags bindingFlags = BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.FlattenHierarchy;
-
-								foreach (var value in methodParams [0].UniqueValues ()) {
-									if (value is SystemTypeValue systemTypeValue) {
-										foreach (var stringParam in methodParams [1].UniqueValues ()) {
-											// TODO: Change this as needed after deciding whether or not we are to keep
-											// all methods on a type that was accessed via reflection.
-											if (stringParam is KnownStringValue stringValue) {
-												MarkMethodsOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, m => m.Name == stringValue.Contents, bindingFlags);
-												reflectionContext.RecordHandledPattern ();
-											} else {
-												RequireDynamicallyAccessedMembers (ref reflectionContext, DynamicallyAccessedMemberKinds.Methods, value, calledMethod.Parameters [0]);
-											}
-										}
-									} else {
-										RequireDynamicallyAccessedMembers (ref reflectionContext, DynamicallyAccessedMemberKinds.Methods, value, calledMethod.Parameters [0]);
-									}
-								}
-							}
-							break;
-
-						//
-						// System.Linq.Expressions.Expression
-						// 
-						// static Field (Expression, Type, String)
-						// static Property (Expression, Type, String)
-						//
-						case var fieldOrPropertyInstrinsic when fieldOrPropertyInstrinsic == IntrinsicId.Expression_Field || fieldOrPropertyInstrinsic == IntrinsicId.Expression_Property: {
-								reflectionContext.AnalyzingPattern ();
-								DynamicallyAccessedMemberKinds memberKind = fieldOrPropertyInstrinsic == IntrinsicId.Expression_Property ? DynamicallyAccessedMemberKinds.Properties : DynamicallyAccessedMemberKinds.Fields;
-
-								foreach (var value in methodParams [1].UniqueValues ()) {
-									if (value is SystemTypeValue systemTypeValue) {
-										foreach (var stringParam in methodParams [2].UniqueValues ()) {
-											if (stringParam is KnownStringValue stringValue) {
-												BindingFlags bindingFlags = methodParams [0].Kind == ValueNodeKind.Null ? BindingFlags.Static : BindingFlags.Default;
-												// TODO: Change this as needed after deciding if we are to keep all fields/properties on a type
-												// that is accessed via reflection. For now, let's only keep the field/property that is retrieved.
-												if (memberKind is DynamicallyAccessedMemberKinds.Properties) {
-													MarkPropertiesOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, filter: p => p.Name == stringValue.Contents, bindingFlags);
-												} else {
-													MarkFieldsOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, filter: f => f.Name == stringValue.Contents, bindingFlags);
-												}
-
-												reflectionContext.RecordHandledPattern ();
-											} else {
-												RequireDynamicallyAccessedMembers (ref reflectionContext, memberKind, value, calledMethod.Parameters [2]);
-											}
-										}
-									} else {
-										RequireDynamicallyAccessedMembers (ref reflectionContext, memberKind, value, calledMethod.Parameters [1]);
-									}
-								}
-							}
-							break;
-
-						//
-						// System.Linq.Expressions.Expression
-						// 
-						// static New (Type)
-						//
-						case IntrinsicId.Expression_New: {
-								reflectionContext.AnalyzingPattern ();
-
-								foreach (var value in methodParams [0].UniqueValues ()) {
-									if (value is SystemTypeValue systemTypeValue) {
-										MarkConstructorsOnType (ref reflectionContext, systemTypeValue.TypeRepresented, null, BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
-										reflectionContext.RecordHandledPattern ();
-									} else {
-										RequireDynamicallyAccessedMembers (ref reflectionContext, DynamicallyAccessedMemberKinds.DefaultConstructor, value, calledMethod.Parameters [0]);
-									}
-								}
-							}
-							break;
-
-						//
-						// System.Type
-						//
-						// GetType (string)
-						// GetType (string, Boolean)
-						// GetType (string, Boolean, Boolean)
-						// GetType (string, Func<AssemblyName, Assembly>, Func<Assembly, String, Boolean, Type>)
-						// GetType (string, Func<AssemblyName, Assembly>, Func<Assembly, String, Boolean, Type>, Boolean)
-						// GetType (string, Func<AssemblyName, Assembly>, Func<Assembly, String, Boolean, Type>, Boolean, Boolean)
-						//
-						case IntrinsicId.Type_GetType: {
-								reflectionContext.AnalyzingPattern ();
-
-								foreach (var typeNameValue in methodParams [0].UniqueValues ()) {
-									if (typeNameValue is KnownStringValue knownStringValue) {
-										TypeDefinition foundType = _markStep.ResolveFullyQualifiedTypeName (knownStringValue.Contents);
-										if (foundType == null) {
-											// Intentionally ignore - it's not wrong for code to call Type.GetType on non-existing name, the code might expect null/exception back.
-											reflectionContext.RecordHandledPattern ();
-										} else {
-											var methodCalling = reflectionContext.MethodCalling;
-											reflectionContext.RecordRecognizedPattern (foundType, () => _markStep.MarkType (foundType, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling)));
-											methodReturnValue = MergePointValue.MergeValues (methodReturnValue, new SystemTypeValue (foundType));
-										}
-									} else if (typeNameValue == NullValue.Instance) {
-										reflectionContext.RecordHandledPattern ();
-									} else if (typeNameValue is LeafValueWithDynamicallyAccessedMemberNode valueWithDynamicallyAccessedMember && valueWithDynamicallyAccessedMember.DynamicallyAccessedMemberKinds != 0) {
-										// Propagate the annotation from the type name to the return value. Annotation on a string value will be fullfilled whenever a value is assigned to the string with annotation.
-										// So while we don't know which type it is, we can guarantee that it will fullfill the annotation.
-										reflectionContext.RecordHandledPattern ();
-										methodReturnValue = MergePointValue.MergeValues (methodReturnValue, new MethodReturnValue (valueWithDynamicallyAccessedMember.DynamicallyAccessedMemberKinds) {
-											SourceContext = calledMethod.Parameters [0]
-										});
-									} else {
-										reflectionContext.RecordUnrecognizedPattern ($"Reflection call '{calledMethod.FullName}' inside '{reflectionContext.MethodCalling.FullName}' was detected with unknown value for the type name.");
-									}
-								}
-
-							}
-							break;
-
-						//
-						// GetConstructor (Type[])
-						// GetConstructor (BindingFlags, Binder, Type[], ParameterModifier [])
-						// GetConstructor (BindingFlags, Binder, CallingConventions, Type[], ParameterModifier [])
-						//
-						case IntrinsicId.Type_GetConstructor: {
-								reflectionContext.AnalyzingPattern ();
-
-								var parameters = calledMethod.Parameters;
-								BindingFlags bindingFlags = BindingFlags.Default;
-								if (parameters.Count > 1) {
-									if (methodParams [1].AsConstInt () != null)
-										bindingFlags |= (BindingFlags)methodParams [1].AsConstInt ();
-								}
-								// Go over all types we've seen
-								foreach (var value in methodParams [0].UniqueValues ()) {
-									if (value is SystemTypeValue systemTypeValue) {
-										MarkConstructorsOnType (ref reflectionContext, systemTypeValue.TypeRepresented, (Func<MethodDefinition, bool>)null, bindingFlags);
-										reflectionContext.RecordHandledPattern ();
-									} else {
-										// Otherwise fall back to the bitfield requirements
-										var requiredMemberKinds = ((bindingFlags & BindingFlags.NonPublic) == 0)
-												? DynamicallyAccessedMemberKinds.PublicConstructors
-												: DynamicallyAccessedMemberKinds.Constructors;
-										RequireDynamicallyAccessedMembers (ref reflectionContext, requiredMemberKinds, value, calledMethod.Parameters [0]);
-									}
-								}
-							}
-							break;
-
-						//
-						// GetMethod (string)
-						// GetMethod (string, BindingFlags)
-						// GetMethod (string, Type[])
-						// GetMethod (string, Type[], ParameterModifier[])
-						// GetMethod (string, BindingFlags, Binder, Type[], ParameterModifier[]) 6
-						// GetMethod (string, BindingFlags, Binder, CallingConventions, Type[], ParameterModifier[]) 7
-						// GetMethod (string, int, Type[])
-						// GetMethod (string, int, Type[], ParameterModifier[]?)
-						// GetMethod (string, int, BindingFlags, Binder?, Type[], ParameterModifier[]?)
-						// GetMethod (string, int, BindingFlags, Binder?, CallingConventions, Type[], ParameterModifier[]?)
-						//
-						case IntrinsicId.Type_GetMethod: {
-								reflectionContext.AnalyzingPattern ();
-
-								BindingFlags bindingFlags = BindingFlags.Default;
-								if (calledMethod.Parameters.Count > 1 && calledMethod.Parameters [1].ParameterType.Name == "BindingFlags" && methodParams [2].AsConstInt () != null) {
-									bindingFlags |= (BindingFlags)methodParams [2].AsConstInt ();
-								} else if (calledMethod.Parameters.Count > 2 && calledMethod.Parameters [2].ParameterType.Name == "BindingFlags" && methodParams [3].AsConstInt () != null) {
-									bindingFlags |= (BindingFlags)methodParams [3].AsConstInt ();
-								}
-
-								foreach (var value in methodParams [0].UniqueValues ()) {
-									if (value is SystemTypeValue systemTypeValue) {
-										foreach (var stringParam in methodParams [1].UniqueValues ()) {
-											if (stringParam is KnownStringValue stringValue) {
-												MarkMethodsOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, m => m.Name == stringValue.Contents, bindingFlags);
-												reflectionContext.RecordHandledPattern ();
-											} else {
-												// Otherwise fall back to the bitfield requirements
-												var requiredMemberKinds = ((bindingFlags & BindingFlags.NonPublic) == 0)
-														? DynamicallyAccessedMemberKinds.PublicMethods
-														: DynamicallyAccessedMemberKinds.Methods;
-												RequireDynamicallyAccessedMembers (ref reflectionContext, requiredMemberKinds, value, calledMethod.Parameters [0]);
-											}
-										}
-									} else {
-										// Otherwise fall back to the bitfield requirements
-										var requiredMemberKinds = ((bindingFlags & BindingFlags.NonPublic) == 0)
-												? DynamicallyAccessedMemberKinds.PublicMethods
-												: DynamicallyAccessedMemberKinds.Methods;
-										RequireDynamicallyAccessedMembers (ref reflectionContext, requiredMemberKinds, value, calledMethod.Parameters [0]);
-									}
-								}
-							}
-							break;
-
-						//
-						// GetField (string)
-						// GetField (string, BindingFlags)
-						// GetEvent (string)
-						// GetEvent (string, BindingFlags)
-						// GetProperty (string)
-						// GetProperty (string, BindingFlags)
-						// GetProperty (string, Type)
-						// GetProperty (string, Type[])
-						// GetProperty (string, Type, Type[])
-						// GetProperty (string, Type, Type[], ParameterModifier[])
-						// GetProperty (string, BindingFlags, Binder, Type, Type[], ParameterModifier[])
-						//
-						case var fieldPropertyOrEvent when ((fieldPropertyOrEvent == IntrinsicId.Type_GetField || fieldPropertyOrEvent == IntrinsicId.Type_GetProperty || fieldPropertyOrEvent == IntrinsicId.Type_GetEvent)
-							&& calledMethod.DeclaringType.Namespace == "System"
-							&& calledMethod.DeclaringType.Name == "Type"
-							&& calledMethod.Parameters [0].ParameterType.FullName == "System.String")
-							&& calledMethod.HasThis: {
-
-								reflectionContext.AnalyzingPattern ();
-								DynamicallyAccessedMemberKinds? memberKind = null;
-								switch (fieldPropertyOrEvent) {
-									case IntrinsicId.Type_GetEvent:
-										memberKind = DynamicallyAccessedMemberKinds.Events;
-										break;
-
-									case IntrinsicId.Type_GetField:
-										memberKind = DynamicallyAccessedMemberKinds.Fields;
-										break;
-
-									case IntrinsicId.Type_GetProperty:
-										memberKind = DynamicallyAccessedMemberKinds.Properties;
-										break;
-
-									default:
-										throw new ArgumentException ($"Reflection call '{calledMethod.FullName}' inside '{callingMethodBody.Method.FullName}' is of unexpected member type.");
-								}
-								if (memberKind == null)
-									break;
-
-								BindingFlags bindingFlags = BindingFlags.Default;
-								if (calledMethod.Parameters.Count > 1 && calledMethod.Parameters [1].ParameterType.Name == "BindingFlags" && methodParams [2].AsConstInt () != null) {
-									bindingFlags |= (BindingFlags)methodParams [2].AsConstInt ();
-								}
-
-								foreach (var value in methodParams [0].UniqueValues ()) {
-									if (value is SystemTypeValue systemTypeValue) {
-										foreach (var stringParam in methodParams [1].UniqueValues ()) {
-											if (stringParam is KnownStringValue stringValue) {
-												switch (memberKind) {
-													case DynamicallyAccessedMemberKinds.Events:
-														MarkEventsOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, filter: e => e.Name == stringValue.Contents, bindingFlags);
-														break;
-													case DynamicallyAccessedMemberKinds.Fields:
-														MarkFieldsOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, filter: f => f.Name == stringValue.Contents, bindingFlags);
-														break;
-													case DynamicallyAccessedMemberKinds.Properties:
-														MarkPropertiesOnTypeHierarchy (ref reflectionContext, systemTypeValue.TypeRepresented, filter: p => p.Name == stringValue.Contents, bindingFlags);
-														break;
-													default:
-														Debug.Fail ("Unreachable.");
-														break;
-												}
-												reflectionContext.RecordHandledPattern ();
-											} else {
-												RequireDynamicallyAccessedMembers (ref reflectionContext, (DynamicallyAccessedMemberKinds)memberKind, value, calledMethod.Parameters [0]);
-											}
-										}
-									} else {
-										RequireDynamicallyAccessedMembers (ref reflectionContext, (DynamicallyAccessedMemberKinds)memberKind, value, calledMethod.Parameters [0]);
-									}
-								}
-							}
-							break;
-
-						//
-						// System.Activator
-						// 
-						// static CreateInstance (System.Type type)
-						// static CreateInstance (System.Type type, bool nonPublic)
-						// static CreateInstance (System.Type type, params object?[]? args)
-						// static CreateInstance (System.Type type, object?[]? args, object?[]? activationAttributes)
-						// static CreateInstance (System.Type type, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object?[]? args, System.Globalization.CultureInfo? culture)
-						// static CreateInstance (System.Type type, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object?[]? args, System.Globalization.CultureInfo? culture, object?[]? activationAttributes) { throw null; }
-						//
-						case IntrinsicId.Activator_CreateInstance_Type: {
-								var parameters = calledMethod.Parameters;
-
-								reflectionContext.AnalyzingPattern ();
-
-								int? ctorParameterCount = null;
-								BindingFlags bindingFlags = BindingFlags.Instance;
-								if (parameters.Count > 1) {
-									if (parameters [1].ParameterType.MetadataType == MetadataType.Boolean) {
-										// The overload that takes a "nonPublic" bool
-										bool nonPublic = true;
-										if (methodParams [1] is ConstIntValue constInt) {
-											nonPublic = constInt.Value != 0;
-										}
-
-										if (nonPublic)
-											bindingFlags |= BindingFlags.NonPublic | BindingFlags.Public;
-										else
-											bindingFlags |= BindingFlags.Public;
-										ctorParameterCount = 0;
-									} else {
-										// Overload that has the parameters as the second or fourth argument
-										int argsParam = parameters.Count == 2 || parameters.Count == 3 ? 1 : 3;
-
-										if (methodParams.Count > argsParam &&
-											methodParams [argsParam] is ArrayValue arrayValue &&
-											arrayValue.Size.AsConstInt () != null) {
-											ctorParameterCount = arrayValue.Size.AsConstInt ();
-										}
-
-										if (parameters.Count > 3) {
-											if (methodParams [1].AsConstInt () != null)
-												bindingFlags |= (BindingFlags)methodParams [1].AsConstInt ();
-											else
-												bindingFlags |= BindingFlags.NonPublic | BindingFlags.Public;
-										} else {
-											bindingFlags |= BindingFlags.Public;
-										}
-									}
-								} else {
-									// The overload with a single System.Type argument
-									ctorParameterCount = 0;
-									bindingFlags |= BindingFlags.Public;
-								}
-
-								// Go over all types we've seen
-								foreach (var value in methodParams [0].UniqueValues ()) {
-									if (value is SystemTypeValue systemTypeValue) {
-										// Special case known type values as we can do better by applying exact binding flags and parameter count.
-										MarkConstructorsOnType (ref reflectionContext, systemTypeValue.TypeRepresented,
-											ctorParameterCount == null ? (Func<MethodDefinition, bool>)null : m => m.Parameters.Count == ctorParameterCount, bindingFlags);
-										reflectionContext.RecordHandledPattern ();
-									} else {
-										// Otherwise fall back to the bitfield requirements
-										var requiredMemberKinds = ctorParameterCount == 0
-											? DynamicallyAccessedMemberKinds.DefaultConstructor
-											: ((bindingFlags & BindingFlags.NonPublic) == 0)
-												? DynamicallyAccessedMemberKinds.PublicConstructors
-												: DynamicallyAccessedMemberKinds.Constructors;
-										RequireDynamicallyAccessedMembers (ref reflectionContext, requiredMemberKinds, value, calledMethod.Parameters [0]);
-									}
-								}
-							}
-							break;
-
-						//
-						// System.Activator
-						// 
-						// static CreateInstance (string assemblyName, string typeName)
-						// static CreateInstance (string assemblyName, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object?[]? args, System.Globalization.CultureInfo? culture, object?[]? activationAttributes)
-						// static CreateInstance (string assemblyName, string typeName, object?[]? activationAttributes)
-						//
-						case IntrinsicId.Activator_CreateInstance_AssemblyName_TypeName:
-							ProcessCreateInstanceByName (ref reflectionContext, calledMethodDefinition, methodParams);
-							break;
-
-						//
-						// System.Activator
-						// 
-						// static CreateInstanceFrom (string assemblyFile, string typeName)
-						// static CreateInstanceFrom (string assemblyFile, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object? []? args, System.Globalization.CultureInfo? culture, object? []? activationAttributes)
-						// static CreateInstanceFrom (string assemblyFile, string typeName, object? []? activationAttributes)
-						//
-						case IntrinsicId.Activator_CreateInstanceFrom:
-							ProcessCreateInstanceByName (ref reflectionContext, calledMethodDefinition, methodParams);
-							break;
-
-						//
-						// System.Activator
-						// 
-						// static T CreateInstance<T> ()
-						//
-						case IntrinsicId.Activator_CreateInstanceOfT: {
-								reflectionContext.AnalyzingPattern ();
-
-								if (calledMethod is GenericInstanceMethod genericCalledMethod && genericCalledMethod.GenericArguments.Count == 1
-									&& genericCalledMethod.GenericArguments [0] is GenericParameter genericParameter) {
-									if (genericParameter.HasDefaultConstructorConstraint) {
-										// This is safe, the linker would have marked the default .ctor already
-										reflectionContext.RecordHandledPattern ();
-										break;
-									}
-								}
-
-								// Not yet supported in any combination
-								reflectionContext.RecordUnrecognizedPattern ($"Activator call '{reflectionContext.MethodCalled.FullName}' inside '{reflectionContext.MethodCalling.FullName}' is not supported");
-							}
-							break;
-
-						//
-						// System.AppDomain
-						//
-						// CreateInstance (string assemblyName, string typeName)
-						// CreateInstance (string assemblyName, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object? []? args, System.Globalization.CultureInfo? culture, object? []? activationAttributes)
-						// CreateInstance (string assemblyName, string typeName, object? []? activationAttributes)
-						//
-						// CreateInstanceAndUnwrap (string assemblyName, string typeName)
-						// CreateInstanceAndUnwrap (string assemblyName, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object? []? args, System.Globalization.CultureInfo? culture, object? []? activationAttributes)
-						// CreateInstanceAndUnwrap (string assemblyName, string typeName, object? []? activationAttributes)
-						//
-						// CreateInstanceFrom (string assemblyFile, string typeName)
-						// CreateInstanceFrom (string assemblyFile, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object? []? args, System.Globalization.CultureInfo? culture, object? []? activationAttributes)
-						// CreateInstanceFrom (string assemblyFile, string typeName, object? []? activationAttributes)
-						//
-						// CreateInstanceFromAndUnwrap (string assemblyFile, string typeName)
-						// CreateInstanceFromAndUnwrap (string assemblyFile, string typeName, bool ignoreCase, System.Reflection.BindingFlags bindingAttr, System.Reflection.Binder? binder, object? []? args, System.Globalization.CultureInfo? culture, object? []? activationAttributes)
-						// CreateInstanceFromAndUnwrap (string assemblyFile, string typeName, object? []? activationAttributes)
-						//
-						case var appDomainCreateInstance when appDomainCreateInstance == IntrinsicId.AppDomain_CreateInstance
-							|| appDomainCreateInstance == IntrinsicId.AppDomain_CreateInstanceAndUnwrap
-							|| appDomainCreateInstance == IntrinsicId.AppDomain_CreateInstanceFrom
-							|| appDomainCreateInstance == IntrinsicId.AppDomain_CreateInstanceFromAndUnwrap:
-							ProcessCreateInstanceByName (ref reflectionContext, calledMethodDefinition, methodParams);
-							break;
-
-						//
-						// System.Reflection.Assembly
-						//
-						// CreateInstance (string typeName)
-						// CreateInstance (string typeName, bool ignoreCase)
-						// CreateInstance (string typeName, bool ignoreCase, BindingFlags bindingAttr, Binder? binder, object []? args, CultureInfo? culture, object []? activationAttributes)
-						//
-						case IntrinsicId.Assembly_CreateInstance:
-							//
-							// TODO: This could be supported for `this` only calls
-							//
-							reflectionContext.AnalyzingPattern ();
-							reflectionContext.RecordUnrecognizedPattern ($"Activator call '{reflectionContext.MethodCalled.FullName}' inside '{reflectionContext.MethodCalling.FullName}' is not yet supported");
-							break;
-
-						default:
-							if (requiresDataFlowAnalysis) {
-								reflectionContext.AnalyzingPattern ();
-								for (int parameterIndex = 0; parameterIndex < methodParams.Count; parameterIndex++) {
-									var requiredMemberKinds = _flowAnnotations.GetParameterAnnotation (calledMethodDefinition, parameterIndex);
-									if (requiredMemberKinds != 0) {
-										IMetadataTokenProvider targetContext;
-										if (calledMethodDefinition.HasImplicitThis ()) {
-											if (parameterIndex == 0)
-												targetContext = calledMethodDefinition;
-											else
-												targetContext = calledMethodDefinition.Parameters [parameterIndex - 1];
-										} else {
-											targetContext = calledMethodDefinition.Parameters [parameterIndex];
-										}
-
-										RequireDynamicallyAccessedMembers (ref reflectionContext, requiredMemberKinds, methodParams [parameterIndex], targetContext);
-									}
-								}
-
-								reflectionContext.RecordHandledPattern ();
-							}
-
-							// To get good reporting of errors we need to track the origin of the value for all method calls
-							// but except Newobj as those are special.
-							if (calledMethodDefinition.ReturnType.MetadataType != MetadataType.Void) {
-								methodReturnValue = new MethodReturnValue (returnValueDynamicallyAccessedMemberKinds) {
-									SourceContext = calledMethodDefinition
-								};
-
-								return true;
-							}
-
-							return false;
-					}
-				}
-				finally {
-					reflectionContext.Dispose ();
-				}
-
-				// If we get here, we handled this as an intrinsic.  As a convenience, if the code above
-				// didn't set the return value (and the method has a return value), we will set it to be an
-				// unknown value with the return type of the method.
-				if (methodReturnValue == null) {
-					if (calledMethod.ReturnType.MetadataType != MetadataType.Void) {
-						methodReturnValue = new MethodReturnValue (returnValueDynamicallyAccessedMemberKinds) {
-							SourceContext = calledMethodDefinition
-						};
-					}
-				}
-
-				// Validate that the return value has the correct annotations as per the method return value annotations
-				if (returnValueDynamicallyAccessedMemberKinds != 0 && methodReturnValue != null) {
-					if (methodReturnValue is LeafValueWithDynamicallyAccessedMemberNode methodReturnValueWithMemberKinds) {
-						if (!methodReturnValueWithMemberKinds.DynamicallyAccessedMemberKinds.HasFlag (returnValueDynamicallyAccessedMemberKinds))
-							throw new InvalidOperationException ($"Internal linker error: processing of call from {callingMethodBody.Method} to {calledMethod} returned value which is not correctly annotated with the expected dynamic member access kinds.");
-					}
-					else if (methodReturnValue is SystemTypeValue) {
-						// SystemTypeValue can fullfill any requirement, so it's always valid
-						// The requirements will be applied at the point where it's consumed (passed as a method parameter, set as field value, returned from the method)
-					}
-					else {
-						throw new InvalidOperationException ($"Internal linker error: processing of call from {callingMethodBody.Method} to {calledMethod} returned value which is not correctly annotated with the expected dynamic member access kinds.");
-					}
-				}
-
-				return true;
-			}
-
-			void ProcessCreateInstanceByName (ref ReflectionPatternContext reflectionContext, MethodDefinition calledMethod, ValueNodeList methodParams)
-			{
-				reflectionContext.AnalyzingPattern ();
-
-				BindingFlags bindingFlags = BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public;
-				bool parameterlessConstructor = true;
-				if (calledMethod.Parameters.Count == 8 && calledMethod.Parameters [2].ParameterType.MetadataType == MetadataType.Boolean &&
-					methodParams [3].AsConstInt () != null) {
-					parameterlessConstructor = false;
-					bindingFlags = BindingFlags.Instance | (BindingFlags)methodParams [3].AsConstInt ();
-				}
-
-				int methodParamsOffset = calledMethod.HasImplicitThis () ? 1 : 0;
-
-				foreach (var assemblyNameValue in methodParams [methodParamsOffset].UniqueValues ()) {
-					if (assemblyNameValue is KnownStringValue assemblyNameStringValue) {
-						foreach (var typeNameValue in methodParams [methodParamsOffset + 1].UniqueValues ()) {
-							if (typeNameValue is KnownStringValue typeNameStringValue) {
-								var resolvedAssembly = _markStep._context.GetLoadedAssembly (assemblyNameStringValue.Contents);
-								if (resolvedAssembly == null) {
-									reflectionContext.RecordUnrecognizedPattern ($"Activator call '{reflectionContext.MethodCalled.FullName}' inside '{reflectionContext.MethodCalling.FullName}' references assembly '{assemblyNameStringValue.Contents}' which could not be found");
-									continue;
-								}
-
-								var resolvedType = FindType (resolvedAssembly, typeNameStringValue.Contents);
-								if (resolvedType == null) {
-									reflectionContext.RecordUnrecognizedPattern ($"Activator call '{reflectionContext.MethodCalled.FullName}' inside '{reflectionContext.MethodCalling.FullName}' references type '{typeNameStringValue}' in assembly '{resolvedAssembly.FullName}' which could not be found");
-									continue;
-								}
-
-								MarkConstructorsOnType (ref reflectionContext, resolvedType,
-									parameterlessConstructor ? m => m.Parameters.Count == 0 : (Func<MethodDefinition, bool>)null, bindingFlags);
-							} else {
-								reflectionContext.RecordUnrecognizedPattern ($"Activator call '{reflectionContext.MethodCalled.FullName}' inside '{reflectionContext.MethodCalling.FullName}' has unrecognized value for the 'typeName' parameter.");
-							}
-						}
-					} else {
-						reflectionContext.RecordUnrecognizedPattern ($"Activator call '{reflectionContext.MethodCalled.FullName}' inside '{reflectionContext.MethodCalling.FullName}' has unrecognized value for the 'assemblyName' parameter.");
-					}
-				}
-			}
-
-			void RequireDynamicallyAccessedMembers (ref ReflectionPatternContext reflectionContext, DynamicallyAccessedMemberKinds requiredMemberKinds, ValueNode value, IMetadataTokenProvider targetContext)
-			{
-				foreach (var uniqueValue in value.UniqueValues ()) {
-					if (uniqueValue is LeafValueWithDynamicallyAccessedMemberNode valueWithDynamicallyAccessedMember) {
-						if (!valueWithDynamicallyAccessedMember.DynamicallyAccessedMemberKinds.HasFlag (requiredMemberKinds)) {
-							reflectionContext.RecordUnrecognizedPattern ($"The {GetValueDescriptionForErrorMessage (valueWithDynamicallyAccessedMember)} " +
-								$"with dynamically accessed member kinds '{GetDynamicallyAccessedMemberKindsDescription (valueWithDynamicallyAccessedMember.DynamicallyAccessedMemberKinds)}' " +
-								$"is passed into the {GetMetadataTokenDescriptionForErrorMessage (targetContext)} " +
-								$"which requires dynamically accessed member kinds `{GetDynamicallyAccessedMemberKindsDescription (requiredMemberKinds)}`. " +
-								$"To fix this add DynamicallyAccessedMembersAttribute to it and specify at least these member kinds '{GetDynamicallyAccessedMemberKindsDescription (requiredMemberKinds)}'.");
-						} else {
-							reflectionContext.RecordHandledPattern ();
-						}
-					} else if (uniqueValue is SystemTypeValue systemTypeValue) {
-						MarkTypeForDynamicallyAccessedMembers (ref reflectionContext, systemTypeValue.TypeRepresented, requiredMemberKinds);
-					} else if (uniqueValue is KnownStringValue knownStringValue) {
-						TypeDefinition foundType = _markStep.ResolveFullyQualifiedTypeName (knownStringValue.Contents);
-						if (foundType == null) {
-							// Intentionally ignore - it's not wrong for code to call Type.GetType on non-existing name, the code might expect null/exception back.
-							reflectionContext.RecordHandledPattern ();
-						} else {
-							MarkTypeForDynamicallyAccessedMembers (ref reflectionContext, foundType, requiredMemberKinds);
-						}
-					} else if (uniqueValue == NullValue.Instance) {
-						// Ignore - probably unreachable path as it would fail at runtime anyway.
-					} else {
-						reflectionContext.RecordUnrecognizedPattern ($"A {GetValueDescriptionForErrorMessage (uniqueValue)} " +
-							$"is passed into the {GetMetadataTokenDescriptionForErrorMessage (targetContext)} " +
-							$"which requires dynamically accessed member kinds `{GetDynamicallyAccessedMemberKindsDescription (requiredMemberKinds)}`. " +
-							$"It's not possible to guarantee that these requirements are met by the application.");
-					}
-				}
-
-				reflectionContext.RecordHandledPattern ();
-			}
-
-			void MarkTypeForDynamicallyAccessedMembers (ref ReflectionPatternContext reflectionContext, TypeDefinition typeDefinition, DynamicallyAccessedMemberKinds requiredMemberKinds)
-			{
-				// Note that it's important to first test for the widest selector (Constructors > PublicConstructors > DefaultConstructor)
-				// as the wider ones include the narrower ones in the bitfield values.
-				if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.Constructors)) {
-					MarkConstructorsOnType (ref reflectionContext, typeDefinition, filter: null);
-				} else if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicConstructors)) {
-					MarkConstructorsOnType (ref reflectionContext, typeDefinition, filter: null, bindingFlags: BindingFlags.Public);
-				} else if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.DefaultConstructor)) {
-					MarkConstructorsOnType (ref reflectionContext, typeDefinition, filter: m => m.IsPublic && m.Parameters.Count == 0);
-				}
-
-				if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.Methods)) {
-					MarkMethodsOnTypeHierarchy (ref reflectionContext, typeDefinition, filter: null);
-				} else if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicMethods)) {
-					MarkMethodsOnTypeHierarchy (ref reflectionContext, typeDefinition, filter: null, bindingFlags: BindingFlags.Public);
-				}
-
-				if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.Fields)) {
-					MarkFieldsOnTypeHierarchy (ref reflectionContext, typeDefinition, filter: null);
-				} else if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicFields)) {
-					MarkFieldsOnTypeHierarchy (ref reflectionContext, typeDefinition, filter: null, bindingFlags: BindingFlags.Public);
-				}
-
-				if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.NestedTypes)) {
-					MarkNestedTypesOnType (ref reflectionContext, typeDefinition, filter: null);
-				} else if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicNestedTypes)) {
-					MarkNestedTypesOnType (ref reflectionContext, typeDefinition, filter: t => t.IsNestedPublic);
-				}
-
-				if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.Properties)) {
-					MarkPropertiesOnTypeHierarchy (ref reflectionContext, typeDefinition, filter: null);
-				} else if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicProperties)) {
-					MarkPropertiesOnTypeHierarchy (ref reflectionContext, typeDefinition, filter: null, bindingFlags: BindingFlags.Public);
-				}
-
-				if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.Events)) {
-					MarkEventsOnTypeHierarchy (ref reflectionContext, typeDefinition, filter: null);
-				} else if (requiredMemberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicEvents)) {
-					MarkEventsOnTypeHierarchy (ref reflectionContext, typeDefinition, filter: null, bindingFlags: BindingFlags.Public);
-				}
-			}
-
-			void MarkConstructorsOnType (ref ReflectionPatternContext reflectionContext, TypeDefinition type, Func<MethodDefinition, bool> filter, BindingFlags? bindingFlags = null)
-			{
-				foreach (var method in type.Methods) {
-					if (!method.IsConstructor)
-						continue;
-
-					if (filter != null && !filter (method))
-						continue;
-
-					if ((bindingFlags & (BindingFlags.Instance | BindingFlags.Static)) == BindingFlags.Static && !method.IsStatic)
-						continue;
-
-					if ((bindingFlags & (BindingFlags.Instance | BindingFlags.Static)) == BindingFlags.Instance && method.IsStatic)
-						continue;
-
-					if ((bindingFlags & (BindingFlags.Public | BindingFlags.NonPublic)) == BindingFlags.Public && !method.IsPublic)
-						continue;
-
-					if ((bindingFlags & (BindingFlags.Public | BindingFlags.NonPublic)) == BindingFlags.NonPublic && method.IsPublic)
-						continue;
-
-					var methodCalling = reflectionContext.MethodCalling;
-					reflectionContext.RecordRecognizedPattern (method, () => _markStep.MarkIndirectlyCalledMethod (method, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling)));
-				}
-			}
-
-			void MarkMethodsOnTypeHierarchy (ref ReflectionPatternContext reflectionContext, TypeDefinition type, Func<MethodDefinition, bool> filter, BindingFlags? bindingFlags = null)
-			{
-				bool onBaseType = false;
-				while (type != null) {
-					foreach (var method in type.Methods) {
-						// Ignore constructors as those are not considered methods from a reflection's point of view
-						if (method.IsConstructor)
-							continue;
-
-						// Ignore private methods on a base type - those are completely ignored by reflection
-						// (anything private on the base type is not visible via the derived type)
-						if (onBaseType && method.IsPrivate)
-							continue;
-
-						// Note that special methods like property getter/setter, event adder/remover will still get through and will be marked.
-						// This is intentional as reflection treats these as methods as well.
-
-						if (filter != null && !filter (method))
-							continue;
-
-						if ((bindingFlags & (BindingFlags.Instance | BindingFlags.Static)) == BindingFlags.Static && !method.IsStatic)
-							continue;
-
-						if ((bindingFlags & (BindingFlags.Instance | BindingFlags.Static)) == BindingFlags.Instance && method.IsStatic)
-							continue;
-
-						if ((bindingFlags & (BindingFlags.Public | BindingFlags.NonPublic)) == BindingFlags.Public && !method.IsPublic)
-							continue;
-
-						if ((bindingFlags & (BindingFlags.Public | BindingFlags.NonPublic)) == BindingFlags.NonPublic && method.IsPublic)
-							continue;
-
-						var methodCalling = reflectionContext.MethodCalling;
-						reflectionContext.RecordRecognizedPattern (method, () => _markStep.MarkIndirectlyCalledMethod (method, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling)));
-					}
-
-					type = type.BaseType?.Resolve ();
-					onBaseType = true;
-				}
-			}
-
-			void MarkFieldsOnTypeHierarchy (ref ReflectionPatternContext reflectionContext, TypeDefinition type, Func<FieldDefinition, bool> filter, BindingFlags bindingFlags = BindingFlags.Default)
-			{
-				bool onBaseType = false;
-				while (type != null) {
-					foreach (var field in type.Fields) {
-						// Ignore private fields on a base type - those are completely ignored by reflection
-						// (anything private on the base type is not visible via the derived type)
-						if (onBaseType && field.IsPrivate)
-							continue;
-
-						// Note that compiler generated fields backing some properties and events will get through here.
-						// This is intentional as reflection treats these as fields as well.
-
-						if (filter != null && !filter (field))
-							continue;
-
-						if ((bindingFlags & (BindingFlags.Instance | BindingFlags.Static)) == BindingFlags.Static && !field.IsStatic)
-							continue;
-
-						if ((bindingFlags & (BindingFlags.Instance | BindingFlags.Static)) == BindingFlags.Instance && field.IsStatic)
-							continue;
-
-						if ((bindingFlags & (BindingFlags.Public | BindingFlags.NonPublic)) == BindingFlags.Public && !field.IsPublic)
-							continue;
-
-						if ((bindingFlags & (BindingFlags.Public | BindingFlags.NonPublic)) == BindingFlags.NonPublic && field.IsPublic)
-							continue;
-
-						var methodCalling = reflectionContext.MethodCalling;
-						reflectionContext.RecordRecognizedPattern (field, () => _markStep.MarkField (field, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling)));
-					}
-
-					type = type.BaseType?.Resolve ();
-					onBaseType = true;
-				}
-			}
-
-			void MarkNestedTypesOnType (ref ReflectionPatternContext reflectionContext, TypeDefinition type, Func<TypeDefinition, bool> filter)
-			{
-				foreach (var nestedType in type.NestedTypes) {
-					if (filter != null && !filter (nestedType))
-						continue;
-
-					var methodCalling = reflectionContext.MethodCalling;
-					reflectionContext.RecordRecognizedPattern (nestedType, () => _markStep.MarkType (nestedType, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling)));
-				}
-			}
-
-			void MarkPropertiesOnTypeHierarchy (ref ReflectionPatternContext reflectionContext, TypeDefinition type, Func<PropertyDefinition, bool> filter, BindingFlags bindingFlags = BindingFlags.Default)
-			{
-				bool onBaseType = false;
-				while (type != null) {
-					foreach (var property in type.Properties) {
-						// Ignore private properties on a base type - those are completely ignored by reflection
-						// (anything private on the base type is not visible via the derived type)
-						// Note that properties themselves are not actually private, their accessors are
-						if (onBaseType &&
-							(property.GetMethod == null || property.GetMethod.IsPrivate) &&
-							(property.SetMethod == null || property.SetMethod.IsPrivate))
-							continue;
-
-						if (filter != null && !filter (property))
-							continue;
-
-						if ((bindingFlags & (BindingFlags.Instance | BindingFlags.Static)) == BindingFlags.Static) { 
-							if ((property.GetMethod != null) && !property.GetMethod.IsStatic) continue;
-							if ((property.SetMethod != null) && !property.SetMethod.IsStatic) continue;
-						}
-
-						if ((bindingFlags & (BindingFlags.Instance | BindingFlags.Static)) == BindingFlags.Instance) {
-							if ((property.GetMethod != null) && property.GetMethod.IsStatic) continue;
-							if ((property.SetMethod != null) && property.SetMethod.IsStatic) continue;
-						}
-
-						if ((bindingFlags & (BindingFlags.Public | BindingFlags.NonPublic)) == BindingFlags.Public) {
-							if ((property.GetMethod == null || !property.GetMethod.IsPublic)
-								&& (property.SetMethod == null || !property.SetMethod.IsPublic))
-								continue;
-						}
-
-						if ((bindingFlags & (BindingFlags.Public | BindingFlags.NonPublic)) == BindingFlags.NonPublic) {
-							if ((property.GetMethod != null) && property.GetMethod.IsPublic) continue;
-							if ((property.SetMethod != null) && property.SetMethod.IsPublic) continue;
-						}
-
-						var methodCalling = reflectionContext.MethodCalling;
-						reflectionContext.RecordRecognizedPattern (property, () => {
-							// Marking the property itself actually doesn't keep it (it only marks its attributes and records the dependency), we have to mark the methods on it
-							_markStep.MarkProperty (property, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling));
-							// TODO - this is sort of questionable - when somebody asks for a property they probably want to call either get or set
-							// but linker tracks those separately, and so accessing the getter/setter will raise a warning as it's potentially trimmed.
-							// So including them here doesn't actually remove the warning even if the code is written correctly.
-							_markStep.MarkMethodIfNotNull (property.GetMethod, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling));
-							_markStep.MarkMethodIfNotNull (property.SetMethod, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling));
-							_markStep.MarkMethodsIf (property.OtherMethods, m => true, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling));
-						});
-					}
-
-					type = type.BaseType?.Resolve ();
-					onBaseType = true;
-				}
-			}
-
-			void MarkEventsOnTypeHierarchy (ref ReflectionPatternContext reflectionContext, TypeDefinition type, Func<EventDefinition, bool> filter, BindingFlags bindingFlags = BindingFlags.Default)
-			{
-				bool onBaseType = false;
-				while (type != null) {
-					foreach (var @event in type.Events) {
-						// Ignore private properties on a base type - those are completely ignored by reflection
-						// (anything private on the base type is not visible via the derived type)
-						// Note that properties themselves are not actually private, their accessors are
-						if (onBaseType &&
-							(@event.AddMethod == null || @event.AddMethod.IsPrivate) &&
-							(@event.RemoveMethod == null || @event.RemoveMethod.IsPrivate))
-							continue;
-
-						if (filter != null && !filter (@event))
-							continue;
-
-						if ((bindingFlags & (BindingFlags.Instance | BindingFlags.Static)) == BindingFlags.Static) {
-							if ((@event.AddMethod != null) && !@event.AddMethod.IsStatic) continue;
-							if ((@event.RemoveMethod != null) && !@event.RemoveMethod.IsStatic) continue;
-						}
-
-						if ((bindingFlags & (BindingFlags.Instance | BindingFlags.Static)) == BindingFlags.Instance) {
-							if ((@event.AddMethod != null) && @event.AddMethod.IsStatic) continue;
-							if ((@event.RemoveMethod != null) && @event.RemoveMethod.IsStatic) continue;
-						}
-
-						if ((bindingFlags & (BindingFlags.Public | BindingFlags.NonPublic)) == BindingFlags.Public) {
-							if ((@event.AddMethod == null || !@event.AddMethod.IsPublic)
-								&& (@event.RemoveMethod == null || !@event.RemoveMethod.IsPublic)) 
-								continue;
-						}
-
-						if ((bindingFlags & (BindingFlags.Public | BindingFlags.NonPublic)) == BindingFlags.NonPublic) {
-							if ((@event.AddMethod != null) && @event.AddMethod.IsPublic) continue;
-							if ((@event.RemoveMethod != null) && @event.RemoveMethod.IsPublic) continue;
-						}
-
-						var methodCalling = reflectionContext.MethodCalling;
-						reflectionContext.RecordRecognizedPattern (@event, () => {
-							// MarkEvent actually marks the add/remove/invoke methods as well, so no need to mark those explicitly
-							_markStep.MarkEvent (@event, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling));
-							_markStep.MarkMethodsIf (@event.OtherMethods, m => true, new DependencyInfo (DependencyKind.AccessedViaReflection, methodCalling));
-						});
-					}
-
-					type = type.BaseType?.Resolve ();
-					onBaseType = true;
-				}
-			}
-
-			string GetValueDescriptionForErrorMessage (ValueNode value)
-			{
-				switch (value) {
-					case MethodParameterValue methodParameterValue: {
-							if (methodParameterValue.SourceContext is MethodDefinition method) {
-								int declaredParameterIndex;
-								if (method.HasImplicitThis ()) {
-									if (methodParameterValue.ParameterIndex == 0)
-										return GetMetadataTokenDescriptionForErrorMessage (method);
-
-									declaredParameterIndex = methodParameterValue.ParameterIndex - 1;
-								} else
-									declaredParameterIndex = methodParameterValue.ParameterIndex;
-
-								if (declaredParameterIndex >= 0 && declaredParameterIndex < method.Parameters.Count)
-									return GetMetadataTokenDescriptionForErrorMessage (method.Parameters [declaredParameterIndex]);
-							}
-
-							return $"parameter #{methodParameterValue.ParameterIndex} of method '{methodParameterValue.SourceContext}'";
-						}
-
-					case MethodReturnValue methodReturnValue: {
-							if (methodReturnValue.SourceContext is MethodDefinition method) {
-								return GetMetadataTokenDescriptionForErrorMessage (method.MethodReturnType);
-							}
-
-							return "method return value";
-						}
-
-					case LoadFieldValue loadFieldValue:
-						return GetMetadataTokenDescriptionForErrorMessage (loadFieldValue.Field);
-
-					default:
-						return $"value from unknown source";
-				}
-			}
-
-			string GetMetadataTokenDescriptionForErrorMessage (IMetadataTokenProvider targetContext)
-			{
-				switch (targetContext) {
-					case ParameterDefinition parameterDefinition: return $"parameter '{parameterDefinition.Name}' of method '{parameterDefinition.Method}'";
-					case MethodReturnType methodReturnType: return $"return value of method '{methodReturnType.Method}'";
-					case FieldDefinition fieldDefinition: return $"field '{fieldDefinition}'";
-					// MethodDefinition is used to represent the "this" parameter as we don't support annotations on the method itself.
-					case MethodDefinition methodDefinition: return $"implicit 'this' parameter of method '{methodDefinition}'";
-					default: return $"'{targetContext}'";
-				};
-			}
-
-			string GetDynamicallyAccessedMemberKindsDescription (DynamicallyAccessedMemberKinds memberKinds)
-			{
-				var results = new List<DynamicallyAccessedMemberKinds> ();
-				if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.Constructors))
-					results.Add (DynamicallyAccessedMemberKinds.Constructors);
-				else if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicConstructors))
-					results.Add (DynamicallyAccessedMemberKinds.PublicConstructors);
-				else if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.DefaultConstructor))
-					results.Add (DynamicallyAccessedMemberKinds.DefaultConstructor);
-				
-				if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.Methods))
-					results.Add (DynamicallyAccessedMemberKinds.Methods);
-				else if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicMethods))
-					results.Add (DynamicallyAccessedMemberKinds.PublicMethods);
-				
-				if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.Properties))
-					results.Add (DynamicallyAccessedMemberKinds.Properties);
-				else if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicProperties))
-					results.Add (DynamicallyAccessedMemberKinds.PublicProperties);
-				
-				if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.Fields))
-					results.Add (DynamicallyAccessedMemberKinds.Fields);
-				else if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicFields))
-					results.Add (DynamicallyAccessedMemberKinds.PublicFields);
-				
-				if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.Events))
-					results.Add (DynamicallyAccessedMemberKinds.Events);
-				else if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicEvents))
-					results.Add (DynamicallyAccessedMemberKinds.PublicEvents);
-				
-				if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.NestedTypes))
-					results.Add (DynamicallyAccessedMemberKinds.NestedTypes);
-				else if (memberKinds.HasFlag (DynamicallyAccessedMemberKinds.PublicNestedTypes))
-					results.Add (DynamicallyAccessedMemberKinds.PublicNestedTypes);
-
-				if (results.Count == 0)
-					return "None";
-
-				return string.Join (" | ", results.Select (r => r.ToString ()));
-			}
 		}
 	}
